dom/power/test/test_wakelock_not_exposed.html

<!--
    https://bugzilla.mozilla.org/show_bug.cgi?id=963366
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test navigator.requestWakeLock is not exposed to non-B2G platform</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
 <script type="application/javascript">
   ok(navigator.requestWakeLock === undefined,
      "navigator.requestWakeLock is not exposed to non-B2G platform");
 </script>
</body>
</html>
